Python exercise1
Exercise:
A company needs a small library that can be used as part of a larger program to calculate a department and/or manager’s monthly expense allocation based on the number and types of employees reporting to the manager or contained within the department. There are other parts of the system that deal with persistence and other related, but not core, features. This library should focus on solving the problem below using the appropriate design practices for the chosen language.
Specific Requirements
•	Developers warrant an allocation of $1000 each.
•	QA Testers warrant an allocation of $500 each.
•	Managers warrant an allocation of $300 each.
•	Managers can have QA Testers, Developers and other managers report to them.
•	Only managers can have people reporting to them.
•	People will report to only a single direct manager.
•	Departments consist of any kinds of employees.
•	Users of this library should be able to:
o	Determine the monthly expense allocation warranted a manager who has various employee types reporting to him/her. The level of depth of the hierarchy should be flexible ideally.
o	Determine the monthly expense allocation warranted a department with various employee types under it at multiple levels deep just as the Manager can.